# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Firat Api",
        "version": "0.1"
    },
    "paths": {
...
        "/api/v1/internshipdocument/all": {
            "get": {
                "tags": [
                    "Staj Doküman"
                ],
                "description": "Staj Dokümanları getir",
                "operationId": "b73523d12678def3ee6085f8c9c545be",
                "responses": {
                    "200": {
                        "description": "Ok"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            }
        },
        "/api/v1/internshipdocument/get": {
            "get": {
                "tags": [
                    "Staj Doküman"
                ],
                "description": "Staj Doküman id'ye göre getir",
                "operationId": "fb313f422c07907a4127697cdaec45e0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Parametre açıklaması",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            }
        },
        "/api/v1/internshipdocument/create": {
            "post": {
                "tags": [
                    "Staj Doküman"
                ],
                "description": "Staj Dokümanı Oluştur",
                "operationId": "01737caba29f0d8f6cd2fa4843e34faa",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "internship_accept",
                                    "internship_payment_accept",
                                    "insurance_required",
                                    "user_id"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "integer"
                                    },
                                    "insurance_required": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "internship_accept": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "internship_payment_accept": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "İşlem başarılı"
                    },
                    "400": {
                        "description": "Geçersiz istek"
                    },
                    "401": {
                        "description": "Yetkisiz"
                    },
                    "500": {
                        "description": "Sunucu hatası"
                    }
                }
            }
        },
        "/api/v1/internshipdocument/update": {
            "post": {
                "tags": [
                    "Staj Doküman"
                ],
                "description": "Staj Dokümanı Oluştur",
                "operationId": "db1d7a577464af7a88f4ced16628519c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "id",
                                    "user_id",
                                    "internship_accept",
                                    "internship_payment_accept",
                                    "insurance_required"
                                ],
                                "properties": {
                                    "id": {
                                        "type": "integer"
                                    },
                                    "user_id": {
                                        "type": "integer"
                                    },
                                    "insurance_required": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "internship_accept": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "internship_payment_accept": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "İşlem başarılı"
                    },
                    "400": {
                        "description": "Geçersiz istek"
                    },
                    "401": {
                        "description": "Yetkisiz"
                    },
                    "500": {
                        "description": "Sunucu hatası"
                    }
                }
            }
        },
        "/api/v1/internshipdocument/delete": {
            "post": {
                "tags": [
                    "Staj Doküman"
                ],
                "description": "Staj Dokümanı Sil",
                "operationId": "d147a2036079ba2dd3acffae8c7185ee",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id",
                                    "user_id"
                                ],
                                "properties": {
                                    "id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "İşlem başarılı"
                    },
                    "400": {
                        "description": "Geçersiz istek"
                    },
                    "401": {
                        "description": "Yetkisiz"
                    },
                    "500": {
                        "description": "Sunucu hatası"
                    }
                }
            }
        },
...
    }
}